+++ fireREST/fmc/license/smartlicense/__init__.py
@@ -1,0 +1,9 @@
+from fireREST.defaults import API_RELEASE_720
+from fireREST.fmc import Resource
+
+
+class SmartLicense(Resource):
+    NAMESPACE = 'platform'
+    PATH = '/license/smartlicenses/{uuid}'
+    MINIMUM_VERSION_REQUIRED_CREATE = API_RELEASE_720
+    MINIMUM_VERSION_REQUIRED_GET = API_RELEASE_720