+++ fireREST/fmc/object/anyconnectexternalbrowserpackage/__init__.py
@@ -1,0 +1,10 @@
+from fireREST.defaults import API_RELEASE_720
+from fireREST.fmc import Resource
+
+
+class AnyconnectExternalBrowserPackage(Resource):
+    PATH = '/object/anyconnectexternalbrowserpackages/{uuid}'
+    MINIMUM_VERSION_REQUIRED_CREATE = API_RELEASE_720
+    MINIMUM_VERSION_REQUIRED_GET = API_RELEASE_720
+    MINIMUM_VERSION_REQUIRED_UPDATE = API_RELEASE_720
+    MINIMUM_VERSION_REQUIRED_DELETE = API_RELEASE_720