+++ fireREST/fmc/policy/__init__.py
@@ -5,13 +5,16 @@ from fireREST.fmc.policy.dynamicaccesspolicy import DynamicAccessPolicy
 from fireREST.fmc.policy.filepolicy import FilePolicy
 from fireREST.fmc.policy.ftdnatpolicy import FtdNatPolicy
 from fireREST.fmc.policy.ftds2svpn import FtdS2sVpn
+from fireREST.fmc.policy.healthpolicy import HealthPolicy
 from fireREST.fmc.policy.identitypolicy import IdentityPolicy
 from fireREST.fmc.policy.intrusionpolicy import IntrusionPolicy
 from fireREST.fmc.policy.networkanalysispolicy import NetworkAnalysisPolicy
+from fireREST.fmc.policy.policylock import PolicyLock
 from fireREST.fmc.policy.prefilterpolicy import PrefilterPolicy
 from fireREST.fmc.policy.ravpn import RaVpn
 from fireREST.fmc.policy.snmpalert import SnmpAlert
 from fireREST.fmc.policy.syslogalert import SyslogAlert
+from fireREST.fmc.policy.umbrelladnspolicy import UmbrellaDnsPolicy
 
 
 class Policy:
@@ -22,9 +25,13 @@ class Policy:
         self.filepolicy = FilePolicy(conn)
         self.ftdnatpolicy = FtdNatPolicy(conn)
         self.ftds2svpn = FtdS2sVpn(conn)
+        self.healthpolicy = HealthPolicy(conn)
         self.identitypolicy = IdentityPolicy(conn)
         self.intrusionpolicy = IntrusionPolicy(conn)
         self.networkanalysispolicy = NetworkAnalysisPolicy(conn)
+        self.policylock = PolicyLock(conn)
         self.prefilterpolicy = PrefilterPolicy(conn)
+        self.ravpn = RaVpn(conn)
         self.snmpalert = SnmpAlert(conn)
         self.syslogalert = SyslogAlert(conn)
+        self.umbrelladnspolicy = UmbrellaDnsPolicy(conn)