+++ fireREST/fmc/policy/healthpolicy/__init__.py
@@ -1,0 +1,7 @@
+from fireREST.defaults import API_RELEASE_720
+from fireREST.fmc import Resource
+
+
+class HealthPolicy(Resource):
+    PATH = '/policy/healthpolicies/{uuid}'
+    MINIMUM_VERSION_REQUIRED_GET = API_RELEASE_720