+++ fireREST/fmc/policy/ravpn/ipsecadvancedsettings/__init__.py
@@ -1,0 +1,10 @@
+from fireREST.defaults import API_RELEASE_720
+from fireREST.fmc import ChildResource
+
+
+class IpsecAdvancedSettings(ChildResource):
+    CONTAINER_NAME = 'RaVpn'
+    CONTAINER_PATH = '/policy/ravpns/{uuid}'
+    PATH = '/policy/ravpns/{container_uuid}/ipsecadvancedsettings/{uuid}'
+    MINIMUM_VERSION_REQUIRED_GET = API_RELEASE_720
+    MINIMUM_VERSION_REQUIRED_UPDATE = API_RELEASE_720