+++ fireREST/fmc/troubleshoot/__init__.py
@@ -1,7 +1,9 @@
 from fireREST.fmc import Connection
 from fireREST.fmc.troubleshoot.packettracer import PacketTracer
+from fireREST.fmc.troubleshoot.task import Task
 
 
 class Troubleshoot:
     def __init__(self, conn: Connection):
         self.packettracer = PacketTracer(conn)
+        self.task = Task(conn)